Move zoomToPoint to extension
