+++ Source/ViewableController.swift
@@ -195,37 +195,11 @@ class ViewableController: UIViewController {
     func doubleTapAction(recognizer: UITapGestureRecognizer) {
         let zoomScale = self.zoomingScrollView.zoomScale == 1 ? self.maxZoomScale() : 1
         
-        let touchPoint = recognizer.location(in: self.zoomingScrollView)
-        
-        print("Touch (\(touchPoint.x), \(touchPoint.y))")
+        let touchPoint = recognizer.location(in: recognizer.view!)
         
         //self.zoomingScrollView.contentOffset = touchPoint
         //self.zoomingScrollView.setZoomScale(zoomScale, animated: true)
-        zoomToPoint(zoomPoint: touchPoint, withScale: zoomScale, animated: true)
-    }
-    
-    func zoomToPoint(zoomPoint: CGPoint, withScale scale: CGFloat, animated: Bool) {
-        
-        //Normalize current content size back to content scale of 1.0f
-        let contentSize = CGSize(width: self.zoomingScrollView.contentSize.width / self.zoomingScrollView.zoomScale,
-                                 height: self.zoomingScrollView.contentSize.height / self.zoomingScrollView.zoomScale)
-        
-        //translate the zoom point to relative to the content rect
-        let relativeZoomPoint = CGPoint(x: (zoomPoint.x / self.zoomingScrollView.bounds.size.width) * contentSize.width,
-                                        y: (zoomPoint.y / self.zoomingScrollView.bounds.size.height) * contentSize.height)
-        
-        //derive the size of the region to zoom to
-        let zoomSize = CGSize(width: self.zoomingScrollView.bounds.size.width / scale,
-                              height: self.zoomingScrollView.bounds.size.height / scale)
-        
-        //offset the zoom rect so the actual zoom point is in the middle of the rectangle
-        let zoomRect = CGRect(x: relativeZoomPoint.x - zoomSize.width / 2.0,
-                              y: relativeZoomPoint.y - zoomSize.height / 2.0,
-                              width: zoomSize.width,
-                              height: zoomSize.height)
-        
-        //apply the resize
-        self.zoomingScrollView.zoom(to: zoomRect, animated: animated)
+        self.zoomingScrollView.zoomToPoint(zoomPoint: touchPoint, withScale: zoomScale, animated: true)
     }
 
     override func viewWillLayoutSubviews() {
